backend/pyproject.toml
[project]
name = "learnflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "celery>=5.6.2",
    "exa-py>=2.7.0",
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "huggingface-hub>=1.6.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.10",
    "markdown>=3.10.2",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "reddis>=1.0.1",
    "sqlalchemy>=2.0.48",
    "supabase>=2.28.3",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]
